refactor: DRY `preview` and `send`
